aur-fetch: add --results
The format:

 fetch:///path/to/fetched
 clone:///path/to/cloned

allows high-level commands like aur-sync to decide on how to inspect
build files. In particular, this allows to divide the vifm view to

 clone/c
 clone/d
 fetch/a
 fetch/b

if packages "c", "d" were non-existing and cloned, and packages "a", "b"
existing and fetched (with subsequent steps depending on --rebase). This
allows to distinguish "new" packages as a first step towards solving #379.

Change aur-build --results to match this format, that is:

 build:///path/to/package.tar.xz